// src/redux/filter/types.ts
export enum sortPropertyEnum {
	RATING_DESC = 'rating',
	RATING_ASC = '-rating',
	PRICE_DESC = 'price',
	PRICE_ASC = '-price',
	TITLE_DESC = 'title',
	TITLE_ASC = '-title',
}

export type SortType = {
	title: string;
	sortProperty: sortPropertyEnum;
};

export interface FilterSliceState {
	searchValue: string;
	categoryId: number;
	currentPage: number;
	sort: SortType;
}